Cleanup phase II
